Configured lint. refactored files. Renamed codebase folder
