+++ package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "start": "nodemon index.js"
+    "start": "nodemon index.js",
+    "lint": "eslint --fix src/**"
   },
   "keywords": [],
   "author": "",
@@ -16,6 +17,9 @@
     "uuid": "^8.3.2"
   },
   "devDependencies": {
+    "eslint": "^7.27.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-plugin-import": "^2.23.3",
     "nodemon": "^2.0.4"
   }
 }
